+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -3,6 +3,7 @@
 -   Fix a bug in the deprecated callback handlers of Form component. #7356
 -   Fix a bug in the `<DateFilter>` component where values were retained when switching between rules #7423
 -   Add `hidden` legend position to `Chart`. #7378
+-   Update aligning `Table` fields with the fallback on isNumeric. #7431
 
 # 8.0.0